Network can process data. And real fast.

+++ SneknetRacing/AI/RacerSample.cs
@@ -7,27 +7,27 @@ namespace SneknetRacing.AI
     {
         public static int SampleSize = 9;
 
-        public double Speed { get; set; }
-        public double EngineRPM { get; set; }
-        public double SurfaceTypeRL { get; set; }
-        public double SurfaceTypeRR { get; set; }
-        public double SurfaceTypeFL { get; set; }
-        public double SurfaceTypeFR { get; set; }
-        public double LapDistance { get; set; }
-        public double WorldPosX { get; set; }
-        public double WorldPosZ { get; set; }
+        public float Speed { get; set; }
+        public float EngineRPM { get; set; }
+        public float SurfaceTypeRL { get; set; }
+        public float SurfaceTypeRR { get; set; }
+        public float SurfaceTypeFL { get; set; }
+        public float SurfaceTypeFR { get; set; }
+        public float LapDistance { get; set; }
+        public float WorldPosX { get; set; }
+        public float WorldPosZ { get; set; }
         /*
-        public double WorldForwardDirX { get; set; }
-        public double WorldForwardDirZ { get; set; }
-        public double WorldRightDirX { get; set; }
-        public double WorldRightDirZ { get; set; }
-        public double Yaw { get; set; }
-        public double Pitch { get; set; }
-        public double Roll { get; set; }
+        public float WorldForwardDirX { get; set; }
+        public float WorldForwardDirZ { get; set; }
+        public float WorldRightDirX { get; set; }
+        public float WorldRightDirZ { get; set; }
+        public float Yaw { get; set; }
+        public float Pitch { get; set; }
+        public float Roll { get; set; }
         */
-        public double Throttle { get; set; }
-        public double Steer { get; set; }
-        public double CurrentGear { get; set; }
+        public float Throttle { get; set; }
+        public float Steer { get; set; }
+        public float CurrentGear { get; set; }
 
         public RacerSample()
         {
